requirements.txt
pandas==2.3.1
numpy==1.24.4
scikit-learn==1.3.2
scikit-surprise==1.1.4
scipy==1.10.1
flask==3.1.1